+++ test/CMakeLists.txt
@@ -19,9 +19,9 @@ enable_testing()
 
 set(SOURCE_FILES polygon_test.cpp cross_section_test.cpp manifold_test.cpp boolean_test.cpp sdf_test.cpp samples_test.cpp test_main.cpp
         boolean_complex_test.cpp
-        manifold_advanced_test.cpp
         smooth_test.cpp
-        properties_test.cpp)
+        properties_test.cpp
+        edge_test.cpp)
 
 if(MANIFOLD_CBIND AND NOT EMSCRIPTEN)
   list(APPEND SOURCE_FILES manifoldc_test.cpp)
