+++ test/manifold_test.cpp
@@ -22,7 +22,20 @@
 #include "test.h"
 #include "tri_dist.h"
 
+namespace {
+
 using namespace manifold;
+
+template <typename T>
+int NumUnique(const std::vector<T>& in) {
+  std::set<int> unique;
+  for (const T& v : in) {
+    unique.emplace(v);
+  }
+  return unique.size();
+}
+
+}  // namespace
 
 /**
  * This tests that turning a mesh into a manifold and returning it to a mesh
@@ -337,94 +350,205 @@ TEST(Manifold, WarpBatch) {
   EXPECT_EQ(prop1.surfaceArea, prop2.surfaceArea);
 }
 
-glm::vec4 CircularTangent(const glm::vec3& tangent, const glm::vec3& edgeVec) {
-  const glm::vec3 dir = glm::normalize(tangent);
+/**
+ * Testing more advanced Manifold operations.
+ */
 
-  float weight = glm::abs(glm::dot(dir, glm::normalize(edgeVec)));
-  if (weight == 0) {
-    weight = 1;
-  }
-  // Quadratic weighted bezier for circular interpolation
-  const glm::vec4 bz2 =
-      weight * glm::vec4(dir * glm::length(edgeVec) / (2 * weight), 1);
-  // Equivalent cubic weighted bezier
-  const glm::vec4 bz3 = glm::mix(glm::vec4(0, 0, 0, 1), bz2, 2 / 3.0f);
-  // Convert from homogeneous form to geometric form
-  return glm::vec4(glm::vec3(bz3) / bz3.w, bz3.w);
+TEST(Manifold, Transform) {
+  Manifold cube = Manifold::Cube({1, 2, 3});
+  Manifold cube2 = cube;
+  cube = cube.Rotate(30, 40, 50).Scale({6, 5, 4}).Translate({1, 2, 3});
+
+  glm::mat3 rX(1.0f, 0.0f, 0.0f,          //
+               0.0f, cosd(30), sind(30),  //
+               0.0f, -sind(30), cosd(30));
+  glm::mat3 rY(cosd(40), 0.0f, -sind(40),  //
+               0.0f, 1.0f, 0.0f,           //
+               sind(40), 0.0f, cosd(40));
+  glm::mat3 rZ(cosd(50), sind(50), 0.0f,   //
+               -sind(50), cosd(50), 0.0f,  //
+               0.0f, 0.0f, 1.0f);
+  glm::mat3 s = glm::mat3(1.0f);
+  s[0][0] = 6;
+  s[1][1] = 5;
+  s[2][2] = 4;
+  glm::mat4x3 transform = glm::mat4x3(s * rZ * rY * rX);
+  transform[3] = glm::vec3(1, 2, 3);
+  cube2 = cube2.Transform(transform);
+
+  Identical(cube.GetMesh(), cube2.GetMesh());
 }
 
-TEST(Manifold, SmoothTorus) {
-  Mesh torusMesh =
-      Manifold::Revolve(CrossSection::Circle(1, 8).Translate({2, 0}), 6)
-          .GetMesh();
-  const int numTri = torusMesh.triVerts.size();
+TEST(Manifold, Slice) {
+  Manifold cube = Manifold::Cube();
+  CrossSection bottom = cube.Slice();
+  CrossSection top = cube.Slice(1);
+  EXPECT_EQ(bottom.Area(), 1);
+  EXPECT_EQ(top.Area(), 0);
+}
 
-  // Create correct toroidal halfedge tangents - SmoothOut() is too generic to
-  // do this perfectly.
-  torusMesh.halfedgeTangent.resize(3 * numTri);
-  for (int tri = 0; tri < numTri; ++tri) {
-    for (const int i : {0, 1, 2}) {
-      glm::vec4& tangent = torusMesh.halfedgeTangent[3 * tri + i];
-      const glm::vec3 v = torusMesh.vertPos[torusMesh.triVerts[tri][i]];
-      const glm::vec3 edge =
-          torusMesh.vertPos[torusMesh.triVerts[tri][(i + 1) % 3]] - v;
-      if (edge.z == 0) {
-        glm::vec3 tan(v.y, -v.x, 0);
-        tan *= glm::sign(glm::dot(tan, edge));
-        tangent = CircularTangent(tan, edge);
-      } else if (glm::abs(glm::determinant(
-                     glm::mat2(glm::vec2(v), glm::vec2(edge)))) < kTolerance) {
-        const float theta = glm::asin(v.z);
-        glm::vec2 xy(v);
-        const float r = glm::length(xy);
-        xy = xy / r * v.z * (r > 2 ? -1.0f : 1.0f);
-        glm::vec3 tan(xy.x, xy.y, glm::cos(theta));
-        tan *= glm::sign(glm::dot(tan, edge));
-        tangent = CircularTangent(tan, edge);
-      } else {
-        tangent = {0, 0, 0, -1};
-      }
-    }
-  }
-
-  Manifold smooth = Manifold(torusMesh).RefineToLength(0.1).CalculateNormals(0);
-  Mesh out = smooth.GetMesh();
-  for (glm::vec3 v : out.vertPos) {
-    glm::vec3 p(v.x, v.y, 0);
-    p = glm::normalize(p) * 2.0f;
-    float r = glm::length(v - p);
-    ASSERT_NEAR(r, 1, 0.005);
-  }
+TEST(Manifold, MeshRelation) {
+  Mesh gyroidMesh = Gyroid();
+  MeshGL gyroidMeshGL = WithIndexColors(gyroidMesh);
+  Manifold gyroid(gyroidMeshGL);
 
 #ifdef MANIFOLD_EXPORT
-  ExportOptions options2;
-  options2.faceted = false;
-  options2.mat.normalChannels = {3, 4, 5};
-  options2.mat.roughness = 0;
-  if (options.exportModels) ExportMesh("smoothTorus.glb", out, options2);
+  ExportOptions opt;
+  opt.mat.roughness = 1;
+  opt.mat.colorChannels = glm::ivec4(3, 4, 5, -1);
+  if (options.exportModels) ExportMesh("gyroid.glb", gyroid.GetMeshGL(), opt);
+#endif
+
+  RelatedGL(gyroid, {gyroidMeshGL});
+}
+
+TEST(Manifold, MeshRelationTransform) {
+  const Manifold cube = Manifold::Cube();
+  const MeshGL cubeGL = cube.GetMeshGL();
+  const Manifold turned = cube.Rotate(45, 90);
+
+  RelatedGL(turned, {cubeGL});
+}
+
+TEST(Manifold, MeshRelationRefine) {
+  const Mesh in = Csaszar();
+  MeshGL inGL = WithIndexColors(in);
+  Manifold csaszar(inGL);
+
+  RelatedGL(csaszar, {inGL});
+  csaszar = csaszar.RefineToLength(1);
+  ExpectMeshes(csaszar, {{9019, 18038, 3}});
+  RelatedGL(csaszar, {inGL});
+
+#ifdef MANIFOLD_EXPORT
+  ExportOptions opt;
+  opt.mat.roughness = 1;
+  opt.mat.colorChannels = glm::ivec4(3, 4, 5, -1);
+  if (options.exportModels) ExportMesh("csaszar.glb", csaszar.GetMeshGL(), opt);
 #endif
 }
 
-TEST(Manifold, SineSurface) {
-  MeshGL surface = LevelSet(
-      [](glm::vec3 p) {
-        float mid = glm::sin(p.x) + glm::sin(p.y);
-        return (p.z > mid - 0.5 && p.z < mid + 0.5) ? 1 : -1;
-      },
-      {glm::vec3(-2 * glm::pi<float>() + 0.2),
-       glm::vec3(0 * glm::pi<float>() - 0.2)},
-      1);
-  Manifold smoothed = Manifold(surface).SmoothOut(50).Refine(8);
-  auto prop = smoothed.GetProperties();
-  EXPECT_NEAR(prop.volume, 8.08, 0.01);
-  EXPECT_NEAR(prop.surfaceArea, 30.85, 0.01);
-  EXPECT_EQ(smoothed.Genus(), 0);
+TEST(Manifold, MeshGLRoundTrip) {
+  const Manifold cylinder = Manifold::Cylinder(2, 1);
+  ASSERT_GE(cylinder.OriginalID(), 0);
+  MeshGL inGL = cylinder.GetMeshGL();
+  const Manifold cylinder2(inGL);
+  const MeshGL outGL = cylinder2.GetMeshGL();
+
+  ASSERT_EQ(inGL.runOriginalID.size(), 1);
+  ASSERT_EQ(outGL.runOriginalID.size(), 1);
+  ASSERT_EQ(outGL.runOriginalID[0], inGL.runOriginalID[0]);
+
+  RelatedGL(cylinder2, {inGL});
+}
+
+void CheckCube(const MeshGL& cubeSTL) {
+  Manifold cube(cubeSTL);
+  EXPECT_EQ(cube.NumTri(), 12);
+  EXPECT_EQ(cube.NumVert(), 8);
+  EXPECT_EQ(cube.NumPropVert(), 24);
+
+  auto prop = cube.GetProperties();
+  EXPECT_FLOAT_EQ(prop.volume, 1.0f);
+  EXPECT_FLOAT_EQ(prop.surfaceArea, 6.0f);
+}
+
+TEST(Manifold, Merge) {
+  MeshGL cubeSTL = CubeSTL();
+  EXPECT_EQ(cubeSTL.NumTri(), 12);
+  EXPECT_EQ(cubeSTL.NumVert(), 36);
+
+  Manifold cubeBad(cubeSTL);
+  EXPECT_TRUE(cubeBad.IsEmpty());
+  EXPECT_EQ(cubeBad.Status(), Manifold::Error::NotManifold);
+
+  EXPECT_TRUE(cubeSTL.Merge());
+  CheckCube(cubeSTL);
+
+  EXPECT_FALSE(cubeSTL.Merge());
+  EXPECT_EQ(cubeSTL.mergeFromVert.size(), 28);
+  cubeSTL.mergeFromVert.resize(14);
+  cubeSTL.mergeToVert.resize(14);
+
+  EXPECT_TRUE(cubeSTL.Merge());
+  EXPECT_EQ(cubeSTL.mergeFromVert.size(), 28);
+  CheckCube(cubeSTL);
+}
+
+TEST(Manifold, FaceIDRoundTrip) {
+  const Manifold cube = Manifold::Cube();
+  ASSERT_GE(cube.OriginalID(), 0);
+  MeshGL inGL = cube.GetMeshGL();
+  ASSERT_EQ(NumUnique(inGL.faceID), 6);
+  inGL.faceID = {0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11};
+
+  const Manifold cube2(inGL);
+  const MeshGL outGL = cube2.GetMeshGL();
+  ASSERT_EQ(NumUnique(outGL.faceID), 12);
+}
+
+TEST(Manifold, MirrorUnion) {
+  auto a = Manifold::Cube({5., 5., 5.}, true);
+  auto b = a.Translate({2.5, 2.5, 2.5});
+  auto result = a + b + b.Mirror({1, 1, 0});
 
 #ifdef MANIFOLD_EXPORT
-  if (options.exportModels) {
-    ExportOptions options2;
-    // options2.mat.colorChannels = {3, 4, 5, -1};
-    ExportMesh("sinesurface.glb", smoothed.GetMeshGL(), options2);
-  }
+  if (options.exportModels)
+    ExportMesh("manifold_mirror_union.glb", result.GetMesh(), {});
 #endif
+
+  auto vol_a = a.GetProperties().volume;
+  EXPECT_FLOAT_EQ(vol_a * 2.75, result.GetProperties().volume);
+  EXPECT_TRUE(a.Mirror(glm::vec3(0)).IsEmpty());
+}
+
+TEST(Manifold, MirrorUnion2) {
+  auto a = Manifold::Cube();
+  auto result = Manifold::Compose({a.Mirror({1, 0, 0})});
+  EXPECT_TRUE(result.MatchesTriNormals());
+}
+
+TEST(Manifold, Invalid) {
+  auto invalid = Manifold::Error::InvalidConstruction;
+  auto circ = CrossSection::Circle(10.);
+  auto empty_circ = CrossSection::Circle(-2.);
+  auto empty_sq = CrossSection::Square(glm::vec2(0.0f));
+
+  EXPECT_EQ(Manifold::Sphere(0).Status(), invalid);
+  EXPECT_EQ(Manifold::Cylinder(0, 5).Status(), invalid);
+  EXPECT_EQ(Manifold::Cylinder(2, -5).Status(), invalid);
+  EXPECT_EQ(Manifold::Cube(glm::vec3(0.0f)).Status(), invalid);
+  EXPECT_EQ(Manifold::Cube({-1, 1, 1}).Status(), invalid);
+  EXPECT_EQ(Manifold::Extrude(circ, 0.).Status(), invalid);
+  EXPECT_EQ(Manifold::Extrude(empty_circ, 10.).Status(), invalid);
+  EXPECT_EQ(Manifold::Revolve(empty_sq).Status(), invalid);
+}
+
+TEST(Manifold, MultiCompose) {
+  auto part = Manifold::Compose({Manifold::Cube({10, 10, 10})});
+  auto finalAssembly =
+      Manifold::Compose({part, part.Translate({0, 10, 0}),
+                         part.Mirror({1, 0, 0}).Translate({10, 0, 0}),
+                         part.Mirror({1, 0, 0}).Translate({10, 10, 0})});
+  EXPECT_FLOAT_EQ(finalAssembly.GetProperties().volume, 4000);
+}
+
+TEST(Manifold, MergeDegenerates) {
+  MeshGL cube = Manifold::Cube(glm::vec3(1), true).GetMeshGL();
+  MeshGL squash;
+  squash.vertProperties = cube.vertProperties;
+  squash.triVerts = cube.triVerts;
+  // Move one vert to the position of its neighbor and remove one triangle
+  // linking them to break the manifold.
+  squash.vertProperties[squash.vertProperties.size() - 1] *= -1;
+  squash.triVerts.resize(squash.triVerts.size() - 3);
+  // Rotate the degenerate triangle to the middle to catch more problems.
+  std::rotate(squash.triVerts.begin(), squash.triVerts.begin() + 3 * 5,
+              squash.triVerts.end());
+  // Merge should remove the now duplicate vertex.
+  EXPECT_TRUE(squash.Merge());
+  // Manifold should remove the triangle with two references to the same vert.
+  Manifold squashed = Manifold(squash);
+  EXPECT_FALSE(squashed.IsEmpty());
+  EXPECT_EQ(squashed.Status(), Manifold::Error::NoError);
 }
