+++ test/smooth_test.cpp
@@ -1,4 +1,4 @@
-// Copyright 2021 The Manifold Authors.
+// Copyright 2024 The Manifold Authors.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,6 +17,7 @@
 #include "cross_section.h"
 #include "manifold.h"
 #include "samples.h"
+#include "sdf.h"
 #include "test.h"
 #include "tri_dist.h"
 
@@ -218,6 +219,98 @@ TEST(Smooth, Csaszar) {
       }
     }
     ExportMesh("smoothCsaszar.glb", out, options);
+  }
+#endif
+}
+
+glm::vec4 CircularTangent(const glm::vec3& tangent, const glm::vec3& edgeVec) {
+  const glm::vec3 dir = glm::normalize(tangent);
+
+  float weight = glm::abs(glm::dot(dir, glm::normalize(edgeVec)));
+  if (weight == 0) {
+    weight = 1;
+  }
+  // Quadratic weighted bezier for circular interpolation
+  const glm::vec4 bz2 =
+      weight * glm::vec4(dir * glm::length(edgeVec) / (2 * weight), 1);
+  // Equivalent cubic weighted bezier
+  const glm::vec4 bz3 = glm::mix(glm::vec4(0, 0, 0, 1), bz2, 2 / 3.0f);
+  // Convert from homogeneous form to geometric form
+  return glm::vec4(glm::vec3(bz3) / bz3.w, bz3.w);
+}
+
+TEST(Smooth, SmoothTorus) {
+  Mesh torusMesh =
+      Manifold::Revolve(CrossSection::Circle(1, 8).Translate({2, 0}), 6)
+          .GetMesh();
+  const int numTri = torusMesh.triVerts.size();
+
+  // Create correct toroidal halfedge tangents - SmoothOut() is too generic to
+  // do this perfectly.
+  torusMesh.halfedgeTangent.resize(3 * numTri);
+  for (int tri = 0; tri < numTri; ++tri) {
+    for (const int i : {0, 1, 2}) {
+      glm::vec4& tangent = torusMesh.halfedgeTangent[3 * tri + i];
+      const glm::vec3 v = torusMesh.vertPos[torusMesh.triVerts[tri][i]];
+      const glm::vec3 edge =
+          torusMesh.vertPos[torusMesh.triVerts[tri][(i + 1) % 3]] - v;
+      if (edge.z == 0) {
+        glm::vec3 tan(v.y, -v.x, 0);
+        tan *= glm::sign(glm::dot(tan, edge));
+        tangent = CircularTangent(tan, edge);
+      } else if (glm::abs(glm::determinant(
+                     glm::mat2(glm::vec2(v), glm::vec2(edge)))) < kTolerance) {
+        const float theta = glm::asin(v.z);
+        glm::vec2 xy(v);
+        const float r = glm::length(xy);
+        xy = xy / r * v.z * (r > 2 ? -1.0f : 1.0f);
+        glm::vec3 tan(xy.x, xy.y, glm::cos(theta));
+        tan *= glm::sign(glm::dot(tan, edge));
+        tangent = CircularTangent(tan, edge);
+      } else {
+        tangent = {0, 0, 0, -1};
+      }
+    }
+  }
+
+  Manifold smooth = Manifold(torusMesh).RefineToLength(0.1).CalculateNormals(0);
+  Mesh out = smooth.GetMesh();
+  for (glm::vec3 v : out.vertPos) {
+    glm::vec3 p(v.x, v.y, 0);
+    p = glm::normalize(p) * 2.0f;
+    float r = glm::length(v - p);
+    ASSERT_NEAR(r, 1, 0.005);
+  }
+
+#ifdef MANIFOLD_EXPORT
+  ExportOptions options2;
+  options2.faceted = false;
+  options2.mat.normalChannels = {3, 4, 5};
+  options2.mat.roughness = 0;
+  if (options.exportModels) ExportMesh("smoothTorus.glb", out, options2);
+#endif
+}
+
+TEST(Smooth, SineSurface) {
+  MeshGL surface = LevelSet(
+      [](glm::vec3 p) {
+        float mid = glm::sin(p.x) + glm::sin(p.y);
+        return (p.z > mid - 0.5 && p.z < mid + 0.5) ? 1 : -1;
+      },
+      {glm::vec3(-2 * glm::pi<float>() + 0.2),
+       glm::vec3(0 * glm::pi<float>() - 0.2)},
+      1);
+  Manifold smoothed = Manifold(surface).SmoothOut(50).Refine(8);
+  auto prop = smoothed.GetProperties();
+  EXPECT_NEAR(prop.volume, 8.08, 0.01);
+  EXPECT_NEAR(prop.surfaceArea, 30.85, 0.01);
+  EXPECT_EQ(smoothed.Genus(), 0);
+
+#ifdef MANIFOLD_EXPORT
+  if (options.exportModels) {
+    ExportOptions options2;
+    // options2.mat.colorChannels = {3, 4, 5, -1};
+    ExportMesh("sinesurface.glb", smoothed.GetMeshGL(), options2);
   }
 #endif
 }
